change double to BigDecimal in OrderLine and tests

+++ src/main/java/com/gft/orders/domain/model/valueObject/OrderLines.java
@@ -2,6 +2,7 @@ package com.gft.orders.domain.model.valueObject;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -11,5 +12,5 @@ public class OrderLines {
   int quantity;
   Boolean refund;
   Double LineWeight;
-  Double LinePrice;
+  BigDecimal LinePrice;
 }
